pane selector: protects vim pane and defaults to split
If only one pane is in the selected window and it is our Vim session
then it won't allow selecting that pane, instead it defaults to the
split that is configured as `autoSplitPane`.

If two or more panes are available and the first one is Vim then it will allow
selecting any of the other panes or split of any pane (including the vim
pane).

However if our Vim session is in any pane but the first one in that case
then it still allows us to select it which should get fixed.

The options won't include pane 1 in either case.