feat: fallback to pane by index if matching id not found
This behavior is configurable, but set by default.

If the pane matching the id of the pane that was selected as target
cannot be found we try to find a pane that happens to be in the same
session and window as the pane that was destroyed.

If we do find it we _promote_ it to the target pane and thus look for
its id going forward, which means moving it around works as did with the
originally selected pane.
We do notify the user of this happening.