feat: convenience around target selection and more (#3)
* select: defaults to current session + window and non-vim pane on first run

- also fixed display panes to be able to stay longer without capturing
  keys

* chore: cleanup obsolete vtr config

* feat: supporting to create panes

- 1v opens pane after first vertically
- 1h opens pane after first horizontally
- when only one pane is present and it is _our_ vim session it opens a
  new one after it by default either vertically or horizontally
  (configurable)

* pane: ability to split pane before another one

* chore: showing instructions regarding pane selection

* chore: no longer loose focus when creating new pane

* pane: communicating back if we created a new pane or not

* send: optionally wait before sending keys for newly created pane

* chore: adding todos

* feat: TmuxUnselectTarget command

* chore: verify target validity before sending keys

* chore: improve pane created/selected message

* chore: add more todos

* fix: send clear sequence

* feat: repeat last command

* chore: shorter command to send command (ironically)

* feat: activate target window when so configured

* chore: update todos

* sessions: helper methods for new selector implementation

* util: to move list item

* selector: ui.select for sessions and windows

* selector: ui select for pane

- still some issues but should pretty much be done

* todo: adding send-up idea

* api: integrate with new way of selecting a target

* selector: remove old API

* sessions: remove some obsolete methods

* selector: handling nil session/window

* chore: delete obsolete impl

* chore: misc cleanup and notes

* feat: TmuxUp to send 'Up' terminal repeating last command there

As part of this we added opts in order to allow overriding if this 'Up'
should be set to `lastCommand`.
The default for this is configurable via `storeUpCommand`.
Additionally made ensureTarget configurable.

We also marked all methods considered private in the runner with an `_`.

* chore: fix window selection when names clash

We solved this by using window id to key a window in the sessions table.
We also now use the id instead of the name in order to address windows
for targeted commands like `split-window`

* feat: resolve vim path identifiers in command

Added `resolveVimPathIdentifiers` to utils which replaces up to one
occurrence of path identifiers like `%`, `%:p`, `$`, etc.
This function is applied to the cmd before it is sent to tmux.

Things like `cargo test % -- --nocapture` and `est %` work now.

* chore: use gsub to expand path identifier

We also now support up to 5 modifiers to percent, i.e. '%t:p:e'

* pane selector: protects vim pane and defaults to split

If only one pane is in the selected window and it is our Vim session
then it won't allow selecting that pane, instead it defaults to the
split that is configured as `autoSplitPane`.

If two or more panes are available and the first one is Vim then it will allow
selecting any of the other panes or split of any pane (including the vim
pane).

However if our Vim session is in any pane but the first one in that case
then it still allows us to select it which should get fixed.

The options won't include pane 1 in either case.

* selector: fix showin available panes

pane module now returns a list of all pane indexes that are available
b/c our vim session isn't in it.

* tmux: including info of each pane in sessions

* chore: add luacheck and fix warnings

* target: storing all pane info and using pane id only as tmux target

* feat: fallback to pane by index if matching id not found

This behavior is configurable, but set by default.

If the pane matching the id of the pane that was selected as target
cannot be found we try to find a pane that happens to be in the same
session and window as the pane that was destroyed.

If we do find it we _promote_ it to the target pane and thus look for
its id going forward, which means moving it around works as did with the
originally selected pane.
We do notify the user of this happening.

* chore: update todo

* chore: simplify pane by index fallback

Basically we are now just setting the target pane to the one at same
index inside the selector.
The rest of the code remains the same as it will operate on that new
target pane.